Fix `yarn run sync --<project>_ref=` command
Was supposed to be fixed in #229 but lost a few commits. We fix
the snakecasing done by commander.js there but don't set
the ref correctly because the name differs from what is set
by lib/config.js. This adds a parameter to the projects
config to specify what should be used for the command line
arg (defaults to name.replace('-', '_').

Fixes #299