Do not iterate through GRDs for every single script and lib
Anything which required Util was also in turn requiring l10nUtil, which performs computations in the module root and has some expectations as to the state of the source tree. This isn't ideal for scripts which run at any time. Whilst the perfect fix would be exporting this functionality as a function to be run on-demand and not in the module root, the fix in this commit reduces the time this occurs by removing the require from Util.js and therefore every other lib and script