Add options to create_dist to build delta installer
To generate delta installer, pass true to build_delta_installer and
pass folder that includes previous versions chrome.7z file to
last_chrome_installer. This folder should be in out dir.
For example, if we want to delta folder for it, create delta folder
to out/Release/delta and copies chrome.7z to it. and pass "delta"
to last_chrome_installer. During the build, couregette.exe wiil be
copired to that folder.