Fix npm run init to work when using NPM >= 8
From NPM >=6 and as explained in RFC 21[1], npm is removing the
npm_config_* and npm_package_* environment variables from the
context of lifecycle scripts and so, while some of them such
as the npm_package_config_* ones will still work on NPM 7, we
need to adapt to this change or it won't be able to setup the
repository from NPM 8 on.

This change does a small change to make sure that we work on
every possible situation, relying on those environment variables
if still using NPM <= 7, or simply reading the package.json file
as a dictionary when using NPM >= 8.

[1] https://github.com/npm/rfcs/blob/main/implemented/0021-reduce-lifecycle-script-environment.md

Resolves https://github.com/brave/brave-browser/issues/19993