creating custom system prompt
i've written two seperate codes one in html to ask the user to give and submit their prompt and then display the prompt as well, in the js code i have added the addEventListener method, when you click the button, it will check what you typed in the input box. If it’s empty, it’ll show an alert asking you to enter something. If you’ve typed a prompt, it will show it below and log it in the console.